Sort Book List
This commit adds the ability to sort the book list using the select
menu at the top of the page.

1. The select menu's onChange method calls App.setSort
2. App's setSort method creates a sort object which contains the
   primary and secondary sorting method, and writes that to state
3. App's componentDidMount method applies the initial sort,
   before the initial render is triggered.
4. BookList's render method runs the applySort method
5. Which sorts the booklist array before rendering

Since both the sort and the book list live in state any change to one
will trigger the App's render method, ensuring the book list is always
properly sorted.

Fixes #2