Move State for Edit Form to BookForm
This commit fixes a performance bug that was caused by storing state
for the edit form in App, rather than in BookForm. By moving it, we
avoid a situation where the entire book list was refreshing with each
keystroke(!) and keep things more isolated. Now, BookForm is only
initialized when called for, and it primes its own state with the
contents of App.state.selectedBook, but then tracks its own changes.

Fix #15